backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0
alembic>=1.12.1
pydantic>=2.9.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-socketio[asyncio]>=5.10.0
python-multipart>=0.0.6
httpx>=0.25.0

